+++ src/any_value/lazy_clone.rs
@@ -1,20 +1,20 @@
 use std::any::TypeId;
-use crate::any_value::{AnyValue2, AnyValue2Cloneable};
+use crate::any_value::{AnyValue, AnyValueCloneable};
 
-/// Makes AnyValue2Cloneable actually [`Clone`]able.
+/// Makes [`AnyValueCloneable`] actually [`Clone`]able.
 /// Lazy clone on consumption.
-pub struct LazyClone<'a, T: AnyValue2Cloneable>{
+pub struct LazyClone<'a, T: AnyValueCloneable>{
     value: &'a T
 }
 
-impl<'a, T: AnyValue2Cloneable> LazyClone<'a, T>{
+impl<'a, T: AnyValueCloneable> LazyClone<'a, T>{
     #[inline]
     pub fn new(value: &'a T) -> Self{
         Self{value}
     }
 }
 
-impl<'a, T: AnyValue2Cloneable> AnyValue2 for LazyClone<'a, T>{
+impl<'a, T: AnyValueCloneable> AnyValue for LazyClone<'a, T>{
     type Type = T::Type;
 
     #[inline]
@@ -38,14 +38,14 @@ impl<'a, T: AnyValue2Cloneable> AnyValue2 for LazyClone<'a, T>{
     }
 }
 
-impl<'a, T: AnyValue2Cloneable> AnyValue2Cloneable for LazyClone<'a, T>{
+impl<'a, T: AnyValueCloneable> AnyValueCloneable for LazyClone<'a, T>{
     #[inline]
     unsafe fn clone_into(&self, out: *mut u8) {
         self.value.clone_into(out);
     }
 }
 
-impl<'a, T: AnyValue2Cloneable> Clone for LazyClone<'a, T>{
+impl<'a, T: AnyValueCloneable> Clone for LazyClone<'a, T>{
     #[inline]
     fn clone(&self) -> Self {
         Self{value: self.value}