+++ src/any_value/wrapper.rs
@@ -5,7 +5,7 @@ use std::ptr;
 use std::ptr::NonNull;
 use crate::any_value::AnyValue;
 
-/// Helper struct to convert concrete type to [`AnyValue`]
+/// Helper struct to convert concrete type to [`AnyValue`].
 pub struct AnyValueWrapper<T: 'static>{
     value: T
 }
@@ -24,22 +24,20 @@ impl<T: 'static> AnyValue for AnyValueWrapper<T> {
     }
 
     #[inline]
-    fn value_size(&self) -> usize {
+    fn size(&self) -> usize {
         size_of::<T>()
     }
 
     #[inline]
-    fn downcast<U: 'static>(self) -> U {
-        assert_eq!(self.value_typeid(), TypeId::of::<U>());
+    unsafe fn downcast_unchecked<U: 'static>(self) -> U {
         // rust don't see that types are the same after assert.
         let value = ManuallyDrop::new(self.value);
         let ptr = &*value as *const T as *const U;
-        unsafe { ptr::read(ptr) }
+        ptr::read(ptr)
     }
 
     #[inline]
-    unsafe fn consume_bytes<F: FnOnce(NonNull<u8>)>(mut self, f: F) {
-        f(NonNull::new_unchecked(&mut self.value as *mut _  as *mut u8));
-        mem::forget(self.value);
+    fn bytes(&self) -> *const u8 {
+        &self.value as *const _ as *const u8
     }
 }