+++ src/ops/remove.rs
@@ -1,51 +1,75 @@
+use std::mem::size_of;
 use std::marker::PhantomData;
-use std::ptr;
+use std::{mem, ptr};
 use std::ptr::NonNull;
-use crate::any_value::Unknown;
+use crate::{AnyVec, copy_bytes_nonoverlapping};
+use crate::any_value::{AnyValue, AnyValueCloneable, Unknown};
 use crate::any_vec_raw::AnyVecRaw;
+use crate::ops::any_vec_ptr::IAnyVecRawPtr;
 use crate::ops::temp::Operation;
+use crate::traits::Trait;
 
-/// Lazily `remove` element on consumption/drop.
-///
-/// This `struct` is created by [`AnyVec::remove`].
-///
-/// [`AnyVec::remove`]: crate::AnyVec::remove
-pub struct Remove<'a, T: 'static = Unknown>{
-    pub(crate) any_vec: &'a mut AnyVecRaw,
-    pub(crate) index: usize,
-    pub(crate) phantom: PhantomData<&'a mut T>
+pub struct Remove<'a, AnyVecPtr: IAnyVecRawPtr, T: 'static = Unknown>{
+    any_vec_ptr: AnyVecPtr,
+    index: usize,
+    last_index: usize,
+    phantom: PhantomData<&'a mut T>
 }
 
-impl<'a, T: 'static> Operation for Remove<'a, T>{
+impl<'a, AnyVecPtr: IAnyVecRawPtr, T: 'static> Remove<'a, AnyVecPtr, T>{
+    #[inline]
+    pub(crate) fn new(any_vec_ptr: AnyVecPtr, index: usize) -> Self{
+        // 1. mem::forget and element drop panic "safety".
+        let any_vec_raw = unsafe{ any_vec_ptr.any_vec_raw().as_mut() };
+        let last_index = any_vec_raw.len - 1;
+        any_vec_raw.len = index;
+
+        Self{any_vec_ptr, index, last_index, phantom: PhantomData}
+    }
+}
+
+impl<'a, AnyVecPtr: IAnyVecRawPtr, T: 'static> Operation for Remove<'a, AnyVecPtr, T>{
+    type AnyVecPtr = AnyVecPtr;
     type Type = T;
 
     #[inline]
-    fn any_vec(&self) -> &AnyVecRaw {
-        self.any_vec
+    fn any_vec_ptr(&self) -> Self::AnyVecPtr {
+        self.any_vec_ptr
     }
 
     #[inline]
-    unsafe fn consume_bytes<F: FnOnce(NonNull<u8>)>(&mut self, f: F) {
-        // Unknown type
-        // ----------------------------
-        let element_size = self.any_vec.element_layout().size();
-        let element = self.any_vec.mem.as_ptr().add(element_size * self.index);
+    fn bytes(&self) -> *const u8 {
+        unsafe{
+            let any_vec_raw = self.any_vec_ptr.any_vec_raw().as_ref();
+            if !Unknown::is::<T>(){
+                any_vec_raw.downcast_ref_unchecked::<T>().as_slice().get_unchecked(self.index)
+                    as *const T as *const u8
+            } else {
+                any_vec_raw.mem.as_ptr()
+                    .add(any_vec_raw.element_layout().size() * self.index)
+            }
+        }
+    }
 
-        // mem::forget and element drop panic "safety".
-        let last_index = self.any_vec.len - 1;
-        self.any_vec.len = self.index;
-
-        // 1. consume
-        f(NonNull::new_unchecked(element));
+    #[inline]
+    fn consume_op(&mut self) {
+    unsafe{
+        let any_vec_raw = self.any_vec_ptr.any_vec_raw().as_mut();
 
         // 2. shift everything left
-        ptr::copy(
-            element.add(element_size),
-            element,
-            element_size * (last_index - self.index)  // self.any_vec.len - self.index - 1
-        );
+        if !Unknown::is::<T>() {
+            let dst: *mut T = any_vec_raw.downcast_mut_unchecked::<T>().as_mut_slice().get_unchecked_mut(self.index);
+            let src = dst.add(1);
+            ptr::copy(src, dst,self.last_index - self.index);
+        } else {
+            let size = any_vec_raw.element_layout().size();
+            let dst = any_vec_raw.mem.as_ptr().add(size * self.index);
+            let src = dst.add(size);
+            ptr::copy(src, dst,size * (self.last_index - self.index));
+        }
 
         // 3. shrink len `self.any_vec.len -= 1`
-        self.any_vec.len = last_index;
+        any_vec_raw.len = self.last_index;
+    }
     }
 }