+++ src/ops/swap_remove.rs
@@ -1,59 +1,83 @@
+use std::any::TypeId;
 use std::mem::size_of;
 use std::marker::PhantomData;
-use std::ptr;
+use std::{mem, ptr};
 use std::ptr::NonNull;
-use crate::copy_bytes_nonoverlapping;
-use crate::any_value::Unknown;
+use crate::{AnyVec, copy_bytes_nonoverlapping};
+use crate::any_value::{AnyValue, AnyValueCloneable, Unknown};
 use crate::any_vec_raw::AnyVecRaw;
+use crate::ops::any_vec_ptr::IAnyVecRawPtr;
 use crate::ops::temp::Operation;
+use crate::traits::Trait;
 
-/// Lazily `swap_remove` element on consumption/drop.
-///
-/// This `struct` is created by [`AnyVec::swap_remove`].
-///
-/// [`AnyVec::swap_remove`]: crate::AnyVec::swap_remove
-pub struct SwapRemove<'a, T: 'static = Unknown>{
-    pub(crate) any_vec: &'a mut AnyVecRaw,
-    pub(crate) index: usize,
-    pub(crate) phantom: PhantomData<&'a mut T>
+pub struct SwapRemove<'a, AnyVecPtr: IAnyVecRawPtr, T: 'static = Unknown>{
+    any_vec_ptr: AnyVecPtr,
+    element: *mut u8,
+    last_index: usize,
+    phantom: PhantomData<&'a mut T>,
 }
 
-impl<'a, T: 'static> Operation for SwapRemove<'a, T>{
+impl<'a, AnyVecPtr: IAnyVecRawPtr, T: 'static> SwapRemove<'a, AnyVecPtr, T>{
+    #[inline]
+    pub(crate) fn new(any_vec_ptr: AnyVecPtr, index: usize) -> Self{
+        let any_vec = unsafe{ any_vec_ptr.any_vec_raw().as_mut() };
+
+        // 1. mem::forget and element drop panic "safety".
+        let last_index = any_vec.len - 1;
+        any_vec.len = index;
+
+        let element: *mut u8 = unsafe{
+            if !Unknown::is::<T>(){
+                any_vec.downcast_mut_unchecked::<T>().as_mut_slice().get_unchecked_mut(index)
+                    as *mut T as *mut u8
+            } else {
+                any_vec.mem.as_ptr().add(any_vec.element_layout().size() * index)
+            }
+        };
+        Self{ any_vec_ptr, element, last_index, phantom: PhantomData }
+    }
+}
+
+impl<'a, AnyVecPtr: IAnyVecRawPtr, T: 'static> Operation for SwapRemove<'a, AnyVecPtr, T>{
+    type AnyVecPtr = AnyVecPtr;
     type Type = T;
 
     #[inline]
-    fn any_vec(&self) -> &AnyVecRaw {
-        self.any_vec
+    fn any_vec_ptr(&self) -> Self::AnyVecPtr {
+        self.any_vec_ptr
     }
 
     #[inline]
-    unsafe fn consume_bytes<F: FnOnce(NonNull<u8>)>(&mut self, f: F) {
-        // TODO: as separate fn
-        let element_size = if Unknown::is::<T>() {
-            self.any_vec.element_layout().size()
-        } else {
-            size_of::<T>()
-        };
-        let element = self.any_vec.mem.as_ptr().add(element_size * self.index);
+    fn bytes(&self) -> *const u8 {
+        self.element
+    }
 
-        // mem::forget and element drop panic "safety".
-        let last_index = self.any_vec.len - 1;
-        self.any_vec.len = self.index;
-
-        // 1. Consume
-        f(NonNull::new_unchecked(element));
-
+    #[inline]
+    fn consume_op(&mut self) {
+    unsafe{
         // 2. overwrite with last element
-        let last_element = self.any_vec.mem.as_ptr().add(element_size * last_index);
-        if self.index != last_index {
-            if Unknown::is::<T>() {
-                copy_bytes_nonoverlapping(last_element, element, element_size);
+        let any_vec_raw = self.any_vec_ptr.any_vec_raw().as_mut();
+        let last_element =
+            if !Unknown::is::<T>() {
+                any_vec_raw.downcast_ref_unchecked::<T>().as_slice()
+                    .get_unchecked(self.last_index) as *const T as *const u8
             } else {
-                ptr::copy_nonoverlapping(last_element as *const T, element as *mut T, 1);
+                any_vec_raw.mem.as_ptr()
+                    .add(any_vec_raw.element_layout().size() * self.last_index)
+            };
+
+        if self.element as *const u8 != last_element {
+            if !Unknown::is::<T>() {
+                ptr::copy_nonoverlapping
+                    (last_element as *const T, self.element as *mut T, 1);
+            } else {
+                copy_bytes_nonoverlapping
+                    (last_element, self.element, any_vec_raw.element_layout().size());
             }
         }
 
         // 3. shrink len `self.any_vec.len -= 1`
-        self.any_vec.len = last_index
+        any_vec_raw.len = self.last_index;
+    }
     }
 }