+++ src/ops2/mod.rs
@@ -1,7 +1,0 @@
-mod temp;
-mod swap_remove;
-mod remove;
-
-pub use temp::TempValue;
-pub use swap_remove::SwapRemove;
-pub use remove::Remove;