+++ tests/any_vec_typed.rs
@@ -1,4 +1,40 @@
+use itertools::assert_equal;
 use any_vec::AnyVec;
+
+#[test]
+fn remove_test() {
+    let mut any_vec: AnyVec = AnyVec::new::<String>();
+    let mut vec = any_vec.downcast_mut::<String>().unwrap();
+    vec.push(String::from("0"));
+    vec.push(String::from("1"));
+    vec.push(String::from("2"));
+    vec.push(String::from("3"));
+    vec.push(String::from("4"));
+
+    // type erased remove
+    vec.remove(2);
+    assert_equal(vec.as_slice(), &[
+        String::from("0"),
+        String::from("1"),
+        String::from("3"),
+        String::from("4"),
+    ]);
+
+    // remove last
+    vec.remove(3);
+    assert_equal(vec.as_slice(), &[
+        String::from("0"),
+        String::from("1"),
+        String::from("3"),
+    ]);
+
+    // remove first
+    vec.remove(0);
+    assert_equal(vec.as_slice(), &[
+        String::from("1"),
+        String::from("3"),
+    ]);
+}
 
 #[test]
 pub fn downcast_mut_test(){