+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+### Changed
+- `any_value::AnyValue::downcast<T>` now return `Option<T>`
+
 ## 0.6.0
 ### Added
 - `AnyVec` now can be Sync, Send, Clone.